fix: float to int for neoPixel Led

+++ LLCS/actuations.py
@@ -8,5 +8,5 @@ def __init__():
 def actuation(input):
     # navigator.set_pwm_freq_hz(1000)
     navigator.set_pwm_channel_value(PwmChannel.Ch1, input)
-    navigator.set_neopixel([[(2000 - input) / 1600 * 255, 0, 0]])
+    navigator.set_neopixel([[int((2000 - input) / 1600) * 255, 0, 0]])
     navigator.set_pwm_enable(True)
